refactor(latex): move compile and source in new package
